+++ docs/CHANGELOG.md
@@ -15,12 +15,23 @@
 - **Random Exploration Search (`behavior_tree.py`)**
   - Added `ExplorationNode`: If there is no task nearby, agent selects a random location within the task generation range and sets it as the `next_waypoint`, enabling the agent to exploare new tasks while moving to this point.
 
+- **Local Search (`simple.py`)**
+  - Renamed `LocalRandomSearch` to `LocalSearch`, which refers to the decision-making methodology for selecting a task within the agent's `situation_awareness_radius`.
+  - Implemented three modes (`Random`, `MinDist`, `MaxUtil`) within `LocalSearch`.
+    - **`Random` mode**: Selects a task randomly.
+    - **`MinDist` mode**: Selects the task with the shortest distance.
+    - **`MaxUtil` mode**: Selects the task with the highest individual utility. Utilizes `weight_factor_cost` (defined in `config.yaml`) to balance task reward and cost.
+  - To this end, `get_tasks_nearby()` was required to be modified such that it can provide "alive(not `completed`)" local tasks. 
+
+
 ### Changes
 - **Random SearchBehavior Tree (`behavior_tree.py`)**
   - **Updated**: Enhanced `DecisionMakingNode` to complete the decision-making process by setting the `assigned_task_id` and recording the corresponding `next_waypoint` in the blackboard. This change removes the need for the `ConsensusCheckingNode`.
   - **Removed**: `ConsensusCheckingNode`
 - **Behavior Tree Structure (`agent.py`)**
   - Modified to use a Fallback node to sequence Decision Making and Task Execution. If these fail (i.e., decision-making is not successful), the agent performs random movement through the Exploration node.
+- **LocalRandomSearch Algorithm (simple.py)**:
+  - Streamlined core logic for clarity by moving additional functionalities (such as random walk) to `DecisionMakingNode`.  
 
 
 ## Version 1.2.2 (24-07-17)